add inference
