+++ src/pure_gan/train.py
@@ -8,99 +8,10 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch.optim import Adam
+from models import Generator, Discriminator
 from utils import check_and_create_dir
 from mnist_dataloader import check_and_process_dataloader
 import wandb
-
-class Generator(nn.Module):
-    """
-    The Generator model class
-    """
-
-    def __init__(self, latent_dim, img_shape, batch_size, activation="ReLU"):
-        """Constructor
-
-        Args:
-            latent_dim (int): Dimension of latent space
-            img_shape (tuple): Tuple containing shape of the image (channels, width, height)
-            activation (str, optional): Define the activation. Defaults to "ReLU".
-        """
-        super(Generator, self).__init__()
-        self.img_shape = img_shape
-        self.batch_size = batch_size
-        if activation=="ReLU":
-            self.activation = nn.ReLU()
-        elif activation=="LeakyReLU":
-            self.activation = nn.LeakyReLU()
-        else:
-            self.activation = nn.Sigmoid()
-
-        self.l1 = nn.Linear(latent_dim, 128)
-        self.l2 = nn.Linear(128,256)
-        self.n1 = nn.BatchNorm1d(256)
-        self.l3 = nn.Linear(256,1024)
-        self.n2 = nn.BatchNorm1d(1024)
-        self.l4 = nn.Linear(1024,2048)
-        self.n3 = nn.BatchNorm1d(2048)
-        self.l_final = nn.Linear(2048, int(np.prod(img_shape)))
-    
-    def forward(self, latent_space):
-        """_summary_ TODO
-
-        Args:
-            latent_space (_type_): _description_
-        """
-        x = self.l1(latent_space)
-        x = self.l2(x)
-        x = self.n1(x)
-        x = self.l3(x)
-        x = self.n2(x)
-        x = self.l4(x)
-        x = self.n3(x)
-        x = self.l_final(x)
-        x = self.activation(x)
-        return x.reshape(-1, self.img_shape[1], self.img_shape[2])
-
-class Discriminator(nn.Module):
-    """
-    The Discriminator class
-    """
-
-    def __init__(self, img_shape, batch_size,activation="ReLU"):
-        """_summary_ TODO
-
-        Args:
-            img_shape (_type_): _description_
-        """
-        super(Discriminator, self).__init__()
-        self.img_shape = img_shape
-
-        if activation=="ReLU":
-            self.activation = nn.ReLU()
-        elif activation=="LeakyReLU":
-            self.activation = nn.LeakyReLU()
-        else:
-            self.activation = nn.Sigmoid()
-        
-        self.l1 = nn.Linear(int(np.prod(img_shape)), 512)
-        self.l2 = nn.Linear(512,256)
-        self.l3 = nn.Linear(256,64)
-        self.l4 = nn.Linear(64,1)
-        self.l_final = nn.Sigmoid()
-    
-    def forward(self, img):
-        """_summary_ TODO
-
-        Args:
-            img (_type_): _description_
-        """
-        img = img.reshape(-1, self.img_shape[1] * self.img_shape[2])
-        img = self.l1(img)
-        img = self.l2(img)
-        img = self.l3(img)
-        img = self.l4(img)
-        img = self.l_final(img)
-        return img
 
 def train(dataloader, epochs, latent_dim, img_shape, batch_size, learning_rate, output_model_dir):
     """_summary_ TODO
@@ -113,14 +24,12 @@ def train(dataloader, epochs, latent_dim, img_shape, batch_size, learning_rate, 
 
     # Loss and model
     value_function_loss = nn.BCELoss()
-    generator_model = Generator(latent_dim,img_shape,batch_size).to(device)
-    disciminator_model = Discriminator(img_shape,batch_size).to(device)
+    generator_model = Generator(latent_dim,img_shape).to(device)
+    disciminator_model = Discriminator(img_shape).to(device)
 
     g_optimizer = Adam(generator_model.parameters(), lr=learning_rate)
     d_optimizer = Adam(disciminator_model.parameters(), lr=learning_rate)
     
-    generator_model.to(device)
-    disciminator_model.to(device)
     generator_model.train()
     disciminator_model.train()
     for epoch in range(epochs):
@@ -179,7 +88,6 @@ if __name__ == "__main__":
     # Create dir
     datetime_now = datetime.datetime.now().isoformat()
     check_and_create_dir(args.output_model_dir)
-    check_and_create_dir("gen_images")
 
     # Dataloader
     dataloader = check_and_process_dataloader("mnist", IMG_SHAPE, args.batch_size)
